refactor(entities): se movio 'entities' a 'engine/game'
