+++ maps/combined_map.overlay.json
@@ -1103,8 +1103,8 @@
     "wal",
     "wal",
     "wal",
-    "wal",
-    "fl5",
+    "fl0",
+    "fl0",
     "fl0",
     "fl0",
     "fl0",
@@ -1233,7 +1233,7 @@
     "fl0",
     "fl0",
     "fl0",
-    "fl1",
+    "fl0",
     "fl0",
     "fl3",
     "fl3",
@@ -1348,7 +1348,7 @@
     "wal",
     "wal",
     "wal",
-    "fl5",
+    "fl0",
     "fl0",
     "fl0",
     "fl0",
@@ -1470,7 +1470,7 @@
     "wal",
     "wal",
     "fl0",
-    "fl5",
+    "fl0",
     "fl0",
     "fl0",
     "fl0",
@@ -1845,7 +1845,7 @@
     "fl4",
     "fl0",
     "fl1",
-    "fl1",
+    "fl0",
     "fl1",
     "fl1",
     "fl1",
