+++ roguelike_project/engine/game/entities/npc/monster.py
