+++ roguelike_project/engine/game/entities/obstacles/obstacle.py
