+++ roguelike_project/engine/game/entities/player/assets.py
