+++ roguelike_project/engine/game/systems/entity.py
@@ -1,4 +1,4 @@
-from roguelike_project.entities.player.player import Player
+from roguelike_project.engine.game.entities.player.player import Player
 from roguelike_project.loaders.load_obstacles import load_obstacles
 from roguelike_project.loaders.load_enemies import load_enemies
 from roguelike_project.loaders.load_buildings import load_buildings
