+++ roguelike_project/engine/game/systems/multiplayer/multiplayer.py
@@ -1,4 +1,4 @@
-from roguelike_project.entities.remote_player.base import RemotePlayer
+from roguelike_project.engine.game.entities.remote_player.base import RemotePlayer
 
 def render_remote_players(state):
     if hasattr(state, "websocket") and state.websocket and state.websocket_connected:
